move spec to common
