+++ common/spec/builder.go
